refactor api folder
