+++ qiskit_ibm_provider/api/auth.py
@@ -37,3 +37,29 @@ class QuantumAuth(AuthBase):
     def get_headers(self) -> Dict:
         """Return authorization information to be stored in header."""
         return {"X-Access-Token": self.access_token}
+
+
+class CloudAuth(AuthBase):
+    """Attaches IBM Cloud Authentication to the given Request object."""
+
+    def __init__(self, api_key: str, crn: str):
+        self.api_key = api_key
+        self.crn = crn
+
+    def __eq__(self, other: object) -> bool:
+        if isinstance(other, CloudAuth):
+            return all(
+                [
+                    self.api_key == other.api_key,
+                    self.crn == other.crn,
+                ]
+            )
+        return False
+
+    def __call__(self, r: PreparedRequest) -> PreparedRequest:
+        r.headers.update(self.get_headers())
+        return r
+
+    def get_headers(self) -> Dict:
+        """Return authorization information to be stored in header."""
+        return {"Service-CRN": self.crn, "Authorization": f"apikey {self.api_key}"}
