+++ qiskit_ibm_provider/api/client_parameters.py
@@ -12,9 +12,10 @@
 
 """Represent IBM Quantum account client parameters."""
 
-from typing import Dict, Optional, Any
+from typing import Dict, Optional, Any, Union
 
-from ..api.auth import QuantumAuth
+from ..utils import get_runtime_api_base_url
+from ..api.auth import QuantumAuth, CloudAuth
 from ..proxies import ProxyConfiguration
 
 TEMPLATE_IBM_HUBS = "{prefix}/Network/{hub}/Groups/{group}/Projects/{project}"
@@ -27,6 +28,7 @@ class ClientParameters:
     def __init__(
         self,
         token: str,
+        channel: Optional[str] = "ibm_quantum",
         url: str = None,
         instance: Optional[str] = None,
         proxies: Optional[ProxyConfiguration] = None,
@@ -35,6 +37,7 @@ class ClientParameters:
         """ClientParameters constructor.
 
         Args:
+            channel: Channel type. ``ibm_cloud`` or ``ibm_quantum``.
             token: IBM Quantum API token.
             url: IBM Quantum URL (gets replaced with a new-style URL with hub, group, project).
             instance: Service instance to use.
@@ -43,13 +46,21 @@ class ClientParameters:
         """
         self.token = token
         self.instance = instance
+        self.channel = channel
         self.url = url
         self.proxies = proxies
         self.verify = verify
 
-    def get_auth_handler(self) -> QuantumAuth:
+    def get_auth_handler(self) -> Union[CloudAuth, QuantumAuth]:
         """Returns the respective authentication handler."""
+        if self.channel == "ibm_cloud":
+            return CloudAuth(api_key=self.token, crn=self.instance)
+
         return QuantumAuth(access_token=self.token)
+
+    def get_runtime_api_base_url(self) -> str:
+        """Returns the Runtime API base url."""
+        return get_runtime_api_base_url(self.url, self.instance)
 
     def connection_parameters(self) -> Dict[str, Any]:
         """Construct connection related parameters.
