+++ qiskit_ibm_provider/api/clients/runtime.py
@@ -21,7 +21,7 @@ from qiskit_ibm_runtime.api.session import RetrySession
 from .backend import BaseBackendClient
 from ..rest.runtime import Runtime
 from ..client_parameters import ClientParameters
-from ....utils.hgp import from_instance_format
+from ...utils.hgp import from_instance_format
 
 logger = logging.getLogger(__name__)
