--- qiskit_ibm_provider/qiskit_ibm_runtime/api/clients/__init__.py
@@ -1,20 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2021.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""IBM Quantum API clients."""
-
-from .base import BaseClient, WebsocketClientCloseCode
-from .account import AccountClient
-from .auth import AuthClient
-from .version import VersionClient
-from .runtime import RuntimeClient
-from .runtime_ws import RuntimeWebsocketClient
