+++ qiskit_ibm_provider/qiskit_ibm_runtime/hub_group_project.py
@@ -21,9 +21,9 @@ from qiskit_ibm_runtime import (  # pylint: disable=unused-import
     qiskit_runtime_service,
 )
 
-from .api.clients import AccountClient
+from ..api.clients import AccountClient
 from ..utils.backend_decoder import configuration_from_server_data
-from .api.client_parameters import ClientParameters
+from ..api.client_parameters import ClientParameters
 from ..utils.hgp import from_instance_format
 
 logger = logging.getLogger(__name__)
