+++ qiskit_ibm_provider/qiskit_ibm_runtime/qiskit_runtime_service.py
@@ -27,11 +27,11 @@ from qiskit.providers.providerutils import filter_backends
 from qiskit_ibm_runtime import ibm_backend
 from ..accounts import AccountManager, Account, AccountType, ChannelType
 from ..proxies import ProxyConfiguration
-from .api.clients import AuthClient, VersionClient
-from .api.clients.runtime import RuntimeClient
+from ..api.clients import AuthClient, VersionClient
+from ..api.clients.runtime import RuntimeClient
 from .constants import QISKIT_IBM_RUNTIME_API_URL
 from .exceptions import IBMNotAuthorizedError, IBMInputValueError, IBMAccountError
-from .api.exceptions import RequestsApiError
+from ..api.exceptions import RequestsApiError
 
 from .hub_group_project import HubGroupProject  # pylint: disable=cyclic-import
 
@@ -42,7 +42,7 @@ from ..utils.utils import validate_job_tags, validate_runtime_options
 from .runtime_program import ParameterNamespace
 from .program.result_decoder import ResultDecoder
 
-from .api.client_parameters import ClientParameters
+from ..api.client_parameters import ClientParameters
 from .runtime_job import RuntimeJob
 from .runtime_options import RuntimeOptions
