+++ test/unit/mock/fake_provider.py
@@ -53,6 +53,7 @@ class FakeProvider(IBMProvider):
             hgp_name = self.DEFAULT_HGPS[idx]
 
             hgp_params = ClientParameters(
+                channel="ibm_quantum",
                 token="some_token",
                 url="some_url",
                 instance=hgp_name,
